refactor(rl): extract dynamo-rl crate at lib/rl (PR A — pure refactor, no behavior change)
