+++ lib/rl/Cargo.toml
@@ -1,0 +1,27 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+[package]
+name = "dynamo-rl"
+description = "RL admin control plane — handlers, state, fan-out, and HTTP facade for /v1/rl/*"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+# Dependency direction: dynamo-llm -> dynamo-rl -> dynamo-runtime.
+# This crate must NOT depend on dynamo-llm.
+
+[dependencies]
+dynamo-runtime = { workspace = true }
+
+axum         = { workspace = true }
+serde        = { workspace = true }
+serde_json   = { workspace = true }
+tokio        = { workspace = true }
+tracing      = { workspace = true }
+anyhow       = { workspace = true }
+futures      = { workspace = true }
+reqwest      = { workspace = true }
